1341-SPLIT-A-STRING-IN-BALANCED-STRINGS Stats: Time: 0 ms (100%), Space: 38.8 MB (11.82%) - LeetHub

+++ 1341-split-a-string-in-balanced-strings/1341-split-a-string-in-balanced-strings.cs
@@ -1,8 +1,8 @@
 public class Solution {
     public int BalancedStringSplit(string s) {
-        int result = 0;
-        int lCount = 0;
-        int rCount = 0;
+        ushort result = 0;
+        ushort lCount = 0;
+        ushort rCount = 0;
         for(int i = 0; i < s.Length; i++)
         {
             if(s[i] == 'R')
